refactor: move theme context to a separate file
