--- src/context/ThemeContext.tsx
@@ -1,22 +1,0 @@
-import { createContext, type ReactNode } from 'react';
-import { useLocalStorage } from '../hooks/useLocalStorage';
-
-type Theme = {
-  theme: string;
-  setTheme: React.Dispatch<React.SetStateAction<string>>;
-};
-
-export const ThemeContext = createContext<Theme>({
-  theme: 'light',
-  setTheme: () => {},
-});
-
-export function ThemeProvider({ children }: { children: ReactNode }) {
-  const [theme, setTheme] = useLocalStorage('theme', 'light');
-
-  return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-}
